+++ controllers/ReportController.go
@@ -34,7 +34,7 @@ func (this *ReportController) Get() {
 		// "TestSuite",
 	}
 
-	expiration := 60 * 60
+	expiration := int64(60 * 60)
 	spec := redis.DefaultSpec().Db(0)
 	client, err := redis.NewSynchClientWithSpec(spec)
 	if err != nil {
